Fix reloading not working properly.
With an sd-notify service, systemd does not acknowledge the reloading
and will timeout unless MONOTONIC_USEC is passed. From the
systemd.service man page:

    When initiating the reload process the service is expected to reply
    with a notification message via sd_notify(3) that contains the
    "RELOADING=1" field in combination with "MONOTONIC_USEC=" set to
    the current monotonic time (i.e. CLOCK_MONOTONIC in
    clock_gettime(2)) in μs, formatted as decimal string. Once reloading
    is complete another notification message must be sent, containing
    "READY=1".

We call the clock_gettime function directly as go does not provide an
access to the system monotonic clock.

This forces us to move the Reloading util to  notify_linux.